+++ pkg/apis/app/v1alpha1/kogitodataindex_types.go
@@ -24,8 +24,6 @@ const KogitoDataIndexCRDName = "kogitodataindices.app.kiegroup.org"
 // KogitoDataIndexSpec defines the desired state of KogitoDataIndex.
 // +k8s:openapi-gen=true
 type KogitoDataIndexSpec struct {
-	InfinispanMeta    `json:",inline"`
-	KafkaMeta         `json:",inline"`
 	KogitoServiceSpec `json:",inline"`
 	// Add custom validation using kubebuilder tags: https://book-v1.book.kubebuilder.io/beyond_basics/generating_crd.html
 }