+++ pkg/apis/app/v1alpha1/kogitoexplainability_types.go
@@ -25,7 +25,6 @@ const KogitoExplainabilityCRDName = "kogitoexplainabilities.app.kiegroup.org"
 // +k8s:openapi-gen=true
 type KogitoExplainabilitySpec struct {
 	KogitoServiceSpec `json:",inline"`
-	KafkaMeta         `json:",inline"`
 	// Add custom validation using kubebuilder tags: https://book-v1.book.kubebuilder.io/beyond_basics/generating_crd.html
 }