+++ pkg/apis/app/v1alpha1/kogitojobsservice_types.go
@@ -21,8 +21,6 @@ import (
 // KogitoJobsServiceSpec defines the desired state of KogitoJobsService.
 // +k8s:openapi-gen=true
 type KogitoJobsServiceSpec struct {
-	InfinispanMeta    `json:",inline"`
-	KafkaMeta         `json:",inline"`
 	KogitoServiceSpec `json:",inline"`
 	// Add custom validation using kubebuilder tags: https://book-v1.book.kubebuilder.io/beyond_basics/generating_crd.html