+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 - [KOGITO-3270](https://issues.redhat.com/browse/KOGITO-3270) - Operator should support ConfigMap injecting in KogitoServices
 - [KOGITO-2905](https://issues.redhat.com/browse/KOGITO-2905) - Change in image definition in the CRDs
+- [KOGITO-3039](https://issues.redhat.com/browse/KOGITO-3039) - Agnostic infrastructure layer configuration on Operator
 - [KOGITO-2113](https://issues.redhat.com/browse/KOGITO-2113) - Simplify binary build deployment
 ## Bug Fixes